requirements.txt
alpaca-py
python-dotenv
pandas
numpy
pytz